refactor: move scrapting utils to utils package
