+++ packages/utils/src/scraping/index.ts
@@ -3,6 +3,7 @@ export * from './html-transformer'
 export { getJobPostingFromHTML } from './job-posting.utils'
 export { parseLinkedinJobUrl } from './linkedin.utils'
 export * from './queries'
+export * from './scraping.utils'
 import AirbnbListingSchema from './airbnb.schema'
 import { JobPostingSchema } from './job-posting.schema'
