+++ packages/utils/tsconfig.json
@@ -6,8 +6,8 @@
     "sourceMap": true,
     "removeComments": true,
     "target": "ESNext",
-    "module": "Node16",
-    "moduleResolution": "node16",
+    "module": "ESNext",
+    "moduleResolution": "bundler",
     "esModuleInterop": true,
     "resolveJsonModule": true,
     "strict": true,
